backend/information-structuring/requirements.txt
# Information Structuring Service Dependencies

# FastAPI and web framework
fastapi>=0.100.0
uvicorn[standard]>=0.20.0

# Database
sqlalchemy>=2.0.0
alembic>=1.10.0

# HTTP client for API calls
httpx>=0.25.0

# Data validation
pydantic>=2.0.0

# Environment variables
python-dotenv>=1.0.0

# Logging
structlog>=23.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Development
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
